allocation calcs with u64s instead of SlotTrackedValues
